functions/requirements.txt
tensorflow-cpu==2.10.0
numpy==1.23.5
flask==2.0.1
flask-cors==3.0.10 